complete CRD documentation
